feat: Add debug symbols
A Simfony function call f(x) is usually compiled as the following
Simplicity expression, where [x] is the compilation of x, and so on.

comp [x] [f]

However, we can write a different Simplicity target expression that has
the same semantics. The input is paired with a false bit. This is piped
into an assertion that checks if the bit is false. The assertion always
succeeds.

comp (pair false [x]) (assertl (drop [f]) CMR)

The lower expression is a more convoluted way of writing the upper
expression. The benefit of the lower expression is that it includes a
CMR that we can use to inject arbitrary data. We use the hash of the
position of the call inside the original Simfony source code as CMR.
This is the debug symbol.

For now, I insert debug symbols into every call that is fallible.
Any failure on the Simplicity Bit Machine can be associated with a
Simfony call expression via the CMR.

Currently, debug symbols are inserted into every compiled program. We
don't want programs bloated with debug symbols on the blockchain. It is
easy to add a "release" mode next to a "debug" mode to the Simfony
compiler. For brevity, I chose to leave this for later.

Note that I had to compile unwraps with slightly more combinators in
order to fit the shape comp [function args] [function body]. I plan to
revert this change in the "release" mod of the Simfony compiler.