fix: Debug symbols override each other
The debug symbol for `dbg!(x)` used the span of `x`. If `x` was an
expression with its own debug symbol, then there would be a conflict.
For example, `dbg!(jet::bip_0340_verify(...))` would have one debug
symbol instead of two.

An easy fix is to use the whole span of `dbg!(x)` for its debug symbol.
We have to trim the `dbg!(` prefix and `)` suffix for the text of the
debug symbol.